OneDrive/Desktop/deploy/Documents/document_reasoning_assistant/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pymupdf>=1.23.0
python-docx>=0.8.11
psycopg2-binary>=2.9.9
requests>=2.31.0
pinecone>=7.3.0
groq>=0.4.1
